+++ tests/test_visitors/test_ast/test_functions/test_call_context/test_range_len.py
@@ -7,6 +7,21 @@ from wemake_python_styleguide.visitors.ast.functions import (
     WrongFunctionCallContextVisitor,
 )
 
+for_template = """
+def function():
+    for list_item in {0}:
+        ...
+"""
+
+wrong_samples = (
+    'range(len())',
+    'range(len(some))',
+    'range(len([]))',
+    'range(1, len(some))',
+    'range(-1, len(some))',
+    'range(0, len(some), -1)',
+)
+
 
 @pytest.mark.parametrize('code', [
     'range(10)',
@@ -25,9 +40,10 @@ def test_correct_range_len(
     parse_ast_tree,
     code,
     default_options,
+    mode,
 ):
     """Testing that ``range()`` can be used."""
-    tree = parse_ast_tree(code)
+    tree = parse_ast_tree(mode(for_template.format(code)))
 
     visitor = WrongFunctionCallContextVisitor(default_options, tree=tree)
     visitor.run()
@@ -35,24 +51,34 @@ def test_correct_range_len(
     assert_errors(visitor, [])
 
 
-@pytest.mark.parametrize('code', [
-    'range(len())',
-    'range(len(some))',
-    'range(len([]))',
-    'range(1, len(some))',
-    'range(-1, len(some))',
-    'range(0, len(some), -1)',
-])
+@pytest.mark.parametrize('code', wrong_samples)
 def test_range_len(
     assert_errors,
     parse_ast_tree,
     code,
     default_options,
+    mode,
 ):
     """Testing that ``range(len(...))`` cannot be used."""
-    tree = parse_ast_tree(code)
+    tree = parse_ast_tree(mode(for_template.format(code)))
 
     visitor = WrongFunctionCallContextVisitor(default_options, tree=tree)
     visitor.run()
 
     assert_errors(visitor, [ImplicitEnumerateViolation])
+
+
+@pytest.mark.parametrize('code', wrong_samples)
+def test_range_len_outside_for_loop(
+    assert_errors,
+    parse_ast_tree,
+    code,
+    default_options,
+):
+    """Testing that ``range(len(...))`` can be used outside loops."""
+    tree = parse_ast_tree(code)
+
+    visitor = WrongFunctionCallContextVisitor(default_options, tree=tree)
+    visitor.run()
+
+    assert_errors(visitor, [])